Bring version control in-line w/ published v1.0.0
The singularity v1.0.0 published to EverTrue’s Chef Server has
these changes, and should be considered “the” v1.0.0 release.

These appear to be changes that should have come in from the
evertrue/eherot/create_install_and_configure_recipes
branch, but as that branch has been merged in a few times, this
simpler approach to getting things up-to-date was taken.

attn @eherot