Remove subscription to install method for service
Unfortunately, the `maven` resource always reports updated by last
action, so the `singularity` service is restarted on every Chef run.
This is undesirable, and causes instability in the cluster.

We will leave it to operators to manually restart Singularity on new
version installations, and attempt to pull request code to the maven
cookbook to solve for this generally.

Maven code for reference:
https://github.com/chef-cookbooks/maven/blob/v1.3.0/providers/default.rb#L72